+++ deal-consumer/src/main/kotlin/com/chaltteok/consumer/order/service/OrderProcessServiceImpl.kt
@@ -2,45 +2,24 @@ package com.chaltteok.consumer.order.service
 
 import com.chaltteok.consumer.order.service.helper.EventHistoryDuplicateChecker
 import com.chaltteok.consumer.order.service.helper.StockDecrementHelper
-import com.chaltteok.core.domain.DailyStock
-import com.chaltteok.core.domain.Order
-import com.chaltteok.core.domain.OrderItem
-import com.chaltteok.core.domain.Payment
-import com.chaltteok.core.domain.User
-import com.chaltteok.core.domain.enums.OrderStatus
-import com.chaltteok.core.domain.enums.PaymentStatus
-import com.chaltteok.core.event.OrderCompletedEvent
 import com.chaltteok.core.repository.dailystock.DailyStockRepository
-import com.chaltteok.core.repository.eventhistory.EventHistoryRepository
-import com.chaltteok.core.repository.order.OrderRepository
-import com.chaltteok.core.repository.orderitem.OrderItemRepository
-import com.chaltteok.core.repository.payment.PaymentRepository
 import com.chaltteok.core.repository.user.UserRepository
 import io.github.oshai.kotlinlogging.KotlinLogging
-import org.springframework.context.ApplicationEventPublisher
 import org.springframework.orm.ObjectOptimisticLockingFailureException
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-import java.time.format.DateTimeFormatter
 
 private val log = KotlinLogging.logger {}
-private val ORDER_DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
 
 private const val MAX_RETRY = 3
 private const val RETRY_DELAY_MS = 50L
-private const val PAYMENT_METHOD_TIMESALE = "TIMESALE"
 
 @Service
 class OrderProcessServiceImpl(
     private val userRepository: UserRepository,
     private val dailyStockRepository: DailyStockRepository,
-    private val orderRepository: OrderRepository,
-    private val orderItemRepository: OrderItemRepository,
-    private val paymentRepository: PaymentRepository,
-    private val eventHistoryRepository: EventHistoryRepository,
     private val duplicateChecker: EventHistoryDuplicateChecker,
     private val stockDecrementHelper: StockDecrementHelper,
-    private val applicationEventPublisher: ApplicationEventPublisher,
+    private val orderConfirmService: OrderConfirmService,
 ) : OrderProcessService {
 
     override fun processOrder(userId: Long, dailyStockId: Long) {
@@ -74,46 +53,7 @@ class OrderProcessServiceImpl(
             }
         }
 
-        confirmOrder(user, dailyStock)
-    }
-
-    @Transactional
-    internal fun confirmOrder(user: User, dailyStock: DailyStock) {
-        val totalPrice = dailyStock.salePrice
-
-        val order = orderRepository.save(
-            Order(user = user, totalPrice = totalPrice, status = OrderStatus.COMPLETED)
-        )
-        orderItemRepository.save(
-            OrderItem(order = order, product = dailyStock.product, quantity = 1, price = dailyStock.salePrice)
-        )
-        paymentRepository.save(
-            Payment(
-                order = order,
-                amount = totalPrice,
-                status = PaymentStatus.SUCCESS,
-                paymentMethod = PAYMENT_METHOD_TIMESALE
-            )
-        )
-
-        // order 미설정인 EventHistory(DuplicateChecker가 생성)에 order 역참조 backfill
-        eventHistoryRepository.findFirstByUserAndDailyStockAndOrderIsNull(user, dailyStock)?.let {
-            it.order = order
-        }
-
-        // 트랜잭션 커밋 후 이메일·알림·통계를 각 EventListener가 처리
-        applicationEventPublisher.publishEvent(
-            OrderCompletedEvent(
-                orderId = order.id ?: error("Order ID null"),
-                orderNumber = order.orderNumber,
-                userEmail = user.email,
-                userName = user.nickname,
-                productName = dailyStock.product.name,
-                totalAmount = totalPrice.toLong(),
-                orderedAt = order.orderedAt.format(ORDER_DATE_FORMATTER),
-            )
-        )
-
-        log.info { "주문 확정 완료 — orderId=${order.id}, userId=${user.id}, dailyStockId=${dailyStock.id}" }
+        // 별도 빈을 통해 호출 → Spring 프록시 경유 → @Transactional 적용
+        orderConfirmService.confirmOrder(user, dailyStock)
     }
 }
